✨ Add chat component with real-time messaging
- Implemented ChatComponent for real-time messaging
- Integrated Firebase Firestore for message storage
- Added LoadingDots component for loading state indication
- Updated CustomButton to use LoadingDots for loading feedback
- Refactored scrollbar styles for better visibility

This feature allows users to engage in real-time conversations
within the application, enhancing user interaction and experience.

Usage:
```jsx
import { ChatComponent } from './components'

// To use the chat component
<ChatComponent courseId="course123" courseData={courseData} />
