+++ src/components/index.js
@@ -5,3 +5,5 @@ export { default as CustomInteractive } from './CustomInteractive'
 export { default as CustomNavBar } from './CustomNavBar'
 export { default as FAQComponent } from './FAQComponent'
 export { default as CustomInput } from './CustomInput'
+export { default as ChatComponent } from './ChatComponent'
+export { default as LoadingDots } from './LoadingDots'
